refactor: AddressSearchUseCase 삭제 및 LocationManager 싱글톤 생성
- Repository 단에서 CLGeocoder생성만 담당하여 이를 삭제
- CLGeocoder의 초기화를 매번해줘야하는 비용과 싱글톤의 메모리 유지 비용을 트레이드 오프
