--- ChildOfWeather/Data/Repository/DefaultAddressSearchRepository.swift
@@ -1,11 +1,0 @@
-import Foundation
-import CoreLocation
-
-final class DefaultAddressSearchRepository: AddressSearchReopsitory {
-    
-    let service: CLGeocoder
-    
-    init() {
-        self.service = CLGeocoder()
-    }
-}
