--- ChildOfWeather/Domain/RepositoryInterface/LocationSearchRepository.swift
@@ -1,7 +1,0 @@
-import Foundation
-import CoreLocation
-
-protocol AddressSearchReopsitory {
-    
-    var service: CLGeocoder { get }
-}
